+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## next
+
+* Support for "type" volume option (defaults to "classic")
+
 ## v0.4.0
 
 * Switched from virtio to qemu, for OVH compatibility